refactor: Move commands, tasks and exceptions to duke package
