+++ src/main/java/duke/command/AddDeadlineCommand.java
@@ -1,11 +1,11 @@
-package command;
+package duke.command;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 
-import exception.DukeException;
-import task.Deadline;
-import task.TaskList;
+import duke.exception.DukeException;
+import duke.task.Deadline;
+import duke.task.TaskList;
 
 public class AddDeadlineCommand implements Command {
