+++ src/main/java/duke/command/CommandType.java
@@ -1,4 +1,4 @@
-package command;
+package duke.command;
 
 public enum CommandType {
     DEADLINE,
