+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,8 +1,8 @@
-package command;
+package duke.command;
 
-import exception.DukeException;
-import task.Task;
-import task.TaskList;
+import duke.exception.DukeException;
+import duke.task.Task;
+import duke.task.TaskList;
 
 public class DeleteCommand implements Command {
 
@@ -17,7 +17,8 @@ public class DeleteCommand implements Command {
     public void execute(TaskList taskList) {
         // Validate index
         if (index < 0 || index >= taskList.size()) {
-            DukeException error = new DukeException("     OOPS!!! Invalid task number. Please provide a valid task index.");
+            DukeException error = new DukeException(
+                    "     OOPS!!! Invalid task number. Please provide a valid task index.");
             System.out.println(error.getMessage());
             System.out.println("    ____________________________________________________________");
             return;
@@ -37,14 +38,16 @@ public class DeleteCommand implements Command {
     // Static parse method to create DeleteCommand from input
     public static Command parse(String[] parts) {
         if (parts.length < 2) {
-            return new InvalidCommand(new DukeException("     OOPS!!! The 'delete' command requires a task number."));
+            return new InvalidCommand(new DukeException(
+                    "     OOPS!!! The 'delete' command requires a task number."));
         }
 
         try {
             int index = Integer.parseInt(parts[1]) - 1; // Convert to 0-based index
             return new DeleteCommand(index);
         } catch (NumberFormatException e) {
-            return new InvalidCommand(new DukeException("     OOPS!!! Invalid task number. Please provide a valid task index."));
+            return new InvalidCommand(new DukeException(
+                    "     OOPS!!! Invalid task number. Please provide a valid task index."));
         }
     }
 }
