+++ src/main/java/duke/task/Task.java
@@ -1,4 +1,4 @@
-package task;
+package duke.task;
 
 public abstract class Task {
